Prompt to restore the last playback.

+++ SrtVideoPlayer.Shared/Constants/Numbers.cs
@@ -5,6 +5,6 @@
         public const int HistoryLengthDefault = 5;
         public const int FontSizeDefault = 16;
         public const int OffsetDefault = 0;
-        public const int LastPositionDefault = 0;
+        public const double LastPositionDefault = 0;
     }
 }
